Refactor: [BADA-169] map app과 pages로 컨벤션에 맞춰 코드 리팩토링 #96
