+++ src/app/rental/page.tsx
@@ -1,31 +1,7 @@
 'use client';
 
-import { useEffect } from 'react';
+import RentalPage from '@/pages/rental/map/page/RentalPage';
 
-import { renderStoreMarkers } from '@/pages/rental/map/lib/renderStoreMarkers';
-import { useFetchStoresHooks } from '@/pages/rental/map/model/useFetchStoresHooks';
-import { useKakaoMapHooks } from '@/pages/rental/map/model/useKakaoMapHooks';
-
-const RentalPage = () => {
-  const { mapRef, map } = useKakaoMapHooks();
-  const stores = useFetchStoresHooks();
-
-  useEffect(() => {
-    if (!map || stores.isLoading) return;
-    renderStoreMarkers(map, stores.stores);
-  }, [map, stores.stores, stores.isLoading]);
-
-  return (
-    <div
-      ref={mapRef}
-      style={{
-        width: '100%',
-        height: '500px',
-        border: '1px solid #ccc',
-        marginTop: '2rem',
-      }}
-    />
-  );
-};
-
-export default RentalPage;
+export default function Page() {
+  return <RentalPage />;
+}
