+++ src/pages/rental/map/page/RentalPage.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import { useEffect } from 'react';
+
+import { renderStoreMarkers } from '@/pages/rental/map/lib/renderStoreMarkers';
+import { useFetchStoresHooks } from '@/pages/rental/map/model/useFetchStoresHooks';
+import { useKakaoMapHooks } from '@/pages/rental/map/model/useKakaoMapHooks';
+
+const RentalPage = () => {
+  const { mapRef, map } = useKakaoMapHooks();
+  const stores = useFetchStoresHooks();
+
+  useEffect(() => {
+    if (!map || stores.isLoading) return;
+    renderStoreMarkers(map, stores.stores);
+  }, [map, stores.stores, stores.isLoading]);
+
+  return (
+    <div
+      ref={mapRef}
+      style={{
+        width: '100%',
+        height: '500px',
+        border: '1px solid #ccc',
+        marginTop: '2rem',
+      }}
+    />
+  );
+};
+
+export default RentalPage;
